Separate CIM repository from Fake_WBEMConnection
Creates a separate file that is the CIMRepository and contains all of
the Server operation request method responders.  _Fake_Wbemconnection will
continue to contain the direct call methods that are a part of setting
up a mock.

The changes include the following:

1. Move all of the server responder methods the ones that started with
_fake_ to a new class (CIMRepository) and rename them without the _fake_
prefix. This means we now have the same names for the server responder
methods (ex. EnumerateClassNames) as the corresponding client methods.

2. Note that InvokeMethod was not moved because it is not really
a CIM repository responder but a way we could mock InvokeMethod
responses. It remains in Fake_WBEMConnection. This class is in a new
file _cimrepository.py and the move includes all of the support methods
for the server repository handlers (ex. _get_subclasses).

3. Modify _MockMOFWbemConnection so that it calls CIMRepository for at
least some of its requests. This is only partly complete at this point
but calls like it GetClass that originally used conn, not use the
repository.

4. Move the use of the ResolverMixin from Fake_WBEMConnection to
CIMRepository.

5. Create a property namespaces in both Fake_WBEMConnection and
CIMRepository to get the namespaces. Fake_WBEMConnection.namespaces
calls CIMRepository.namespaces which calls the DataStore.namespaces.

6. Put add_namespaces and remove_namespaces in both Fake_WBEMConnection
and CIMRepository with the same hiearchy of calls as 4 above. This
allows each layer its own call to manage namespaces for now. Since the
add and remove of namespaces is currently a public Fake_WBEMConnection
interface we kept those APIs (Note: remove the leanding _ from remove.

7. Modify a number of other methods in Fake_WBEMConnection that called
local methods to call the methods moved to CIMRepository (i.e. from
self.validate_namespace() to self.cimrepository.validate_namespace()

8. Numerous minor changes to the tests to correctly locate the methods
used in various tests.

9. Modify all use of variable qualifier_repo in _resolvermixin to
qualifier_store consistent with other changes made in part 2 pr
